move app.py to root folder
